Moved app code into sub-package to avoid flash dependencies when testing model
